+++ ray_tracer/src/camera.rs
@@ -10,15 +10,15 @@ use crate::{
 pub struct Camera {
     pub hsize: i16,
     pub vsize: i16,
-    pub field_of_view: f32,
-    pub half_width: f32,
-    pub half_height: f32,
-    pub pixel_size: f32,
-    pub transform: Matrix<f32>,
+    pub field_of_view: f64,
+    pub half_width: f64,
+    pub half_height: f64,
+    pub pixel_size: f64,
+    pub transform: Matrix<f64>,
 }
 
 impl Camera {
-    pub fn new(hsize: i16, vsize: i16, field_of_view: f32) -> Self {
+    pub fn new(hsize: i16, vsize: i16, field_of_view: f64) -> Self {
         let (half_width, half_height, pixel_size) =
             Self::compute_fields(hsize, vsize, field_of_view);
 
@@ -34,8 +34,8 @@ impl Camera {
     }
 
     pub fn ray_for_pixel(&self, x: i16, y: i16) -> Ray {
-        let x_offset = (x as f32 + 0.5) * self.pixel_size;
-        let y_offset = (y as f32 + 0.5) * self.pixel_size;
+        let x_offset = (x as f64 + 0.5) * self.pixel_size;
+        let y_offset = (y as f64 + 0.5) * self.pixel_size;
 
         let world_x = self.half_width - x_offset;
         let world_y = self.half_height - y_offset;
@@ -49,9 +49,9 @@ impl Camera {
         return ray(origin, direction);
     }
 
-    fn compute_fields(hsize: i16, vsize: i16, field_of_view: f32) -> (f32, f32, f32) {
+    fn compute_fields(hsize: i16, vsize: i16, field_of_view: f64) -> (f64, f64, f64) {
         let half_view = (field_of_view / 2.).tan();
-        let aspect = (hsize as f32) / (vsize as f32);
+        let aspect = (hsize as f64) / (vsize as f64);
         let half_width;
         let half_height;
         if aspect >= 1. {
@@ -61,13 +61,13 @@ impl Camera {
             half_width = half_view * aspect;
             half_height = half_view;
         }
-        let pixel_size = (half_width * 2.) / (hsize as f32);
+        let pixel_size = (half_width * 2.) / (hsize as f64);
 
         (half_width, half_height, pixel_size)
     }
 }
 
-pub fn view_transform(from: Tuple, to: Tuple, up: Tuple) -> Matrix<f32> {
+pub fn view_transform(from: Tuple, to: Tuple, up: Tuple) -> Matrix<f64> {
     let forward = (to - from).normalize();
     let upn = up.normalize();
     let left = forward.cross(upn);
@@ -83,7 +83,7 @@ pub fn view_transform(from: Tuple, to: Tuple, up: Tuple) -> Matrix<f32> {
     return orientation * translation(-from.x, -from.y, -from.z);
 }
 
-pub fn camera(hsize: i16, vsize: i16, field_of_view: f32) -> Camera {
+pub fn camera(hsize: i16, vsize: i16, field_of_view: f64) -> Camera {
     Camera::new(hsize, vsize, field_of_view)
 }
 
@@ -92,6 +92,7 @@ pub fn render(camera: Camera, world: World) -> Canvas {
     for y in 0..camera.vsize {
         for x in 0..camera.hsize {
             let ray = &camera.ray_for_pixel(x, y);
+
             let color = world.color_at(&ray);
 
             match image.write_pixel(x as usize, y as usize, color) {
